rename button to appbutton for vue multi-word lint rule
